- Restructed Files - Modulerized Scraping Function - Modulerilized Criteria
