+++ .env.example
@@ -1,12 +1,15 @@
 # .env.example - Copy this to .env and fill in with your details
 
-# SMTP Settings
+# SMTP (use your real creds in your private .env, never commit it)
 SMTP_HOST=smtp.gmail.com
 SMTP_PORT=587
 SMTP_USER=your_email@example.com
 SMTP_PASSWORD=your_app_password
 FROM_EMAIL=your_email@example.com
-TO_EMAILS=recipient1@example.com,recipient2@example.com,recipient3@example.com  # Multiple recipients
 
 # Other settings
 LOGGING_LEVEL=DEBUG
+
+# Per-provider recipients (comma-separated)
+RECIPIENTS__TTEC_NORTH_EAST=ops@example.com,me@example.com
+RECIPIENTS__TTEC_EAST_ONLY=alerts@example.com
