+++ .gitignore
@@ -1,3 +1,4 @@
 /.venv
-/__pycache__
+__pycache__/
 .env
+secret
