--- config.yaml
@@ -1,26 +1,0 @@
-# config.yaml
-websites:
-  - url: "https://ttec.co.tt/cis/outages_public.html"
-    area_keywords:
-      - "north"
-      - "east"
-      # - "central"
-      # - "south"
-      # - "west"
-      # - "tobago"
-    location_keywords:
-      - "Perseverance"
-      - "Moka"
-      - "Hummingbird" 
-      - "Fernandez" 
-      - "Saddle" 
-    status_inactive_keyword: "CANCELLED"
-    title: "TTEC"
-    schedule: "0 8 * * 1"  # Cron: every Monday at 8:00 AM
-  # - url: "https://another-outage-site.com"
-  #   area_keywords:
-  #     - "central"
-  #   location_keywords:
-  #     - "san juan"
-  #     - "chaguanas"
-  #   schedule: "0 9 * * 2"  # Cron: every Tuesday at 9:00 AM
