+++ src/email/email_util.py
@@ -1,4 +1,4 @@
-# email_util.py
+# email/email_util.py
 import os, smtplib
 from email.mime.multipart import MIMEMultipart
 from email.mime.text import MIMEText
@@ -12,19 +12,16 @@ SMTP_PORT = int(os.getenv('SMTP_PORT', 587))
 SMTP_USER = os.getenv('SMTP_USER')
 SMTP_PASSWORD = os.getenv('SMTP_PASSWORD')
 FROM_EMAIL = os.getenv('FROM_EMAIL')
-TO_EMAILS_ENV = [e.strip() for e in os.getenv('TO_EMAILS','').split(',') if e.strip()]
 
-def send_email_with_attachment(subject, body_html, attachment_path, recipients=None, logger=None):
-    to_list = recipients if recipients else TO_EMAILS_ENV
-    if not to_list:
-        if logger: logger.error("No recipients configured. Set TO_EMAILS or pass recipients.")
+def send_email_with_attachment(subject, body_html, attachment_path, recipients, logger=None):
+    if not recipients:
+        if logger: logger.error("No recipients provided for this provider.")
         return
-
     msg = MIMEMultipart()
     msg['From'] = FROM_EMAIL
-    msg['To'] = ", ".join(to_list)
+    msg['To'] = ", ".join(recipients)
     msg['Subject'] = subject
-    msg.attach(MIMEText(body_html, 'html'))  # <-- HTML body
+    msg.attach(MIMEText(body_html, 'html'))
 
     if attachment_path:
         with open(attachment_path, "rb") as f:
@@ -38,7 +35,7 @@ def send_email_with_attachment(subject, body_html, attachment_path, recipients=N
         with smtplib.SMTP(SMTP_HOST, SMTP_PORT) as server:
             server.starttls()
             server.login(SMTP_USER, SMTP_PASSWORD)
-            server.sendmail(FROM_EMAIL, to_list, msg.as_string())
-            if logger: logger.info(f"Email sent to {', '.join(to_list)} with attachment {attachment_path}")
+            server.sendmail(FROM_EMAIL, recipients, msg.as_string())
+            if logger: logger.info(f"Email sent to {', '.join(recipients)} with attachment {attachment_path}")
     except Exception as e:
         if logger: logger.error(f"Failed to send email: {str(e)}")
